generate proto file
Assumes --with-tracewrap points to a directory which contains
frame.piqi  This will generate frame.piqi.proto which will call
protoc-c to generate the frame.piqi.pb.c and frame.piqi.pb.h.

This change was added since the old frame.piqi.pb.[ch] files were
contained in the repo and not generated from frame.piqi.  This led
to problems building, since the old files used an out of date version
of protobuf